✅ Subarray Product Less Than K | Medium | Java | Runtime: 0 ms | Memory: 42.6 MB | 2026-08-24

+++ medium/0713-subarray-product-less-than-k/solution.java
@@ -6,7 +6,7 @@ class Solution {
             }
         }
         
-            int prod=1;
+            long prod=1;
                 if(prod<k){
                     count+=j-i+1;
                 }
